refactor: `Header`를 `SummaryHeader`로 변경하여 명확히 한다
