+++ pages/side-panel/src/components/SummaryHeader.tsx
