+++ pages/side-panel/src/components/index.ts
@@ -1,6 +1,6 @@
 export { default as Summary } from './Summary';
 export { default as MemoForm } from './MemoForm';
-export { default as Header } from './Header';
+export { default as SummaryHeader } from './SummaryHeader';
 export { default as SummaryProvider } from './SummaryProvider';
 export { default as LoginSection } from './LoginSection';
 export { default as MemoHeader } from './MemoHeader';
